Use `options` internally to create `autocomplete-option` components
They only receive an `index` and each option has a value that should be displayed, which keeps the internal structure independent of the context they are used in. That paves the way to be able to do a non-block form of the component.

Nomenclature: `items` are the objects that are specific to the context (for example, artists), `options` are the internal objects that items are translated into.